refactor: replace some files into new directory
