+++ src/components/main/forms/UI/Button.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TProps } from '../../../types/props-types';
+import { TProps } from '../../../../types/props-types';
 
 export default function Button(props: TProps) {
   return (
