+++ src/components/main/home/Home.scss
@@ -1,3 +1,5 @@
+@use '../../../styles' as *;
+
 .search {
   flex-basis: 31.25rem;
   flex-shrink: 1;
