+++ src/components/main/home/UI/Cards.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './Cards.scss';
-import { TProps } from '../../../types/props-types';
-import music from '../../../data/data';
+import { TProps } from '../../../../types/props-types';
+import music from '../../../../data/data';
 import CardsInfoRow from './CardsInfo';
 
 export default class Cards extends Component {
